+++ packages/brick_offline_first_with_rest_build/test/offline_first_schema_generator/test_with_associations.dart
@@ -25,7 +25,8 @@ final schema = Schema(0,
           columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true)
-          }),
+          },
+          indices: <SchemaIndex>{}),
       SchemaTable('_brick_OneToOneAssocation_assocs',
           columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
@@ -50,7 +51,8 @@ final schema = Schema(0,
                 foreignTableName: 'SqliteAssoc',
                 onDeleteCascade: false,
                 onDeleteSetDefault: false)
-          })
+          },
+          indices: <SchemaIndex>{})
     });
 ''';