+++ packages/brick_sqlite_abstract/test/db/schema_test.dart
@@ -226,13 +226,15 @@ Schema(
       'demo',
       columns: <SchemaColumn>{
         SchemaColumn('_brick_id', int, autoincrement: true, nullable: false, isPrimaryKey: true)
-      }
+      },
+      indices: <SchemaIndex>{}
     ),
     SchemaTable(
       'demo2',
       columns: <SchemaColumn>{
         SchemaColumn('_brick_id', int, autoincrement: true, nullable: false, isPrimaryKey: true)
-      }
+      },
+      indices: <SchemaIndex>{}
     )
   }
 )''');