+++ packages/brick_sqlite_generators/test/sqlite_schema/test_all_field_types.dart
@@ -27,7 +27,8 @@ final schema = Schema(0,
             SchemaColumn('list', String),
             SchemaColumn('longer_camelized_variable', String),
             SchemaColumn('casing', int)
-          })
+          },
+          indices: <SchemaIndex>{})
     });
 ''';