+++ packages/brick_sqlite_generators/test/sqlite_schema/test_nullable.dart
@@ -19,7 +19,8 @@ final schema = Schema(0,
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true),
             SchemaColumn('name', String, nullable: false)
-          })
+          },
+          indices: <SchemaIndex>{})
     });
 ''';