+++ packages/brick_sqlite_generators/test/sqlite_schema/test_one_to_one_association.dart
@@ -29,7 +29,8 @@ final schema = Schema(0,
           columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true)
-          }),
+          },
+          indices: <SchemaIndex>{}),
       SchemaTable('OneToOneAssocation',
           columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
@@ -44,7 +45,8 @@ final schema = Schema(0,
                 foreignTableName: 'SqliteAssoc',
                 onDeleteCascade: false,
                 onDeleteSetDefault: false)
-          })
+          },
+          indices: <SchemaIndex>{})
     });
 ''';