+++ packages/brick_sqlite_generators/test/sqlite_schema/test_simple.dart
@@ -61,7 +61,8 @@ final schema = Schema(0,
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true),
             SchemaColumn('name', String)
-          })
+          },
+          indices: <SchemaIndex>{})
     });
 ''';